vm: move OpInfo and VmContext from validation module
